refactor window.library to browser.library
