+++ lib/browser.library.ts
